Shape tools refactor
