+++ editor/src/messages/tool/tool_messages/mod.rs
@@ -11,8 +11,8 @@ pub mod navigate_tool;
 pub mod path_tool;
 pub mod pen_tool;
 pub mod polygon_tool;
-pub mod rectangle_tool;
 pub mod select_tool;
+pub mod shape_tool;
 pub mod spline_tool;
 pub mod text_tool;
