+++ editor/src/messages/tool/utility_types.rs
@@ -328,7 +328,7 @@ pub enum ToolType {
 	Freehand,
 	Spline,
 	Line,
-	Rectangle,
+	Shape,
 	Ellipse,
 	Polygon,
 	Text,
@@ -368,7 +368,7 @@ fn list_tools_in_groups() -> Vec<Vec<ToolAvailability>> {
 			ToolAvailability::Available(Box::<freehand_tool::FreehandTool>::default()),
 			ToolAvailability::Available(Box::<spline_tool::SplineTool>::default()),
 			ToolAvailability::Available(Box::<line_tool::LineTool>::default()),
-			ToolAvailability::Available(Box::<rectangle_tool::RectangleTool>::default()),
+			ToolAvailability::Available(Box::<shape_tool::ShapeTool>::default()),
 			ToolAvailability::Available(Box::<ellipse_tool::EllipseTool>::default()),
 			ToolAvailability::Available(Box::<polygon_tool::PolygonTool>::default()),
 			ToolAvailability::Available(Box::<text_tool::TextTool>::default()),
@@ -403,7 +403,7 @@ pub fn tool_message_to_tool_type(tool_message: &ToolMessage) -> ToolType {
 		ToolMessage::Freehand(_) => ToolType::Freehand,
 		ToolMessage::Spline(_) => ToolType::Spline,
 		ToolMessage::Line(_) => ToolType::Line,
-		ToolMessage::Rectangle(_) => ToolType::Rectangle,
+		ToolMessage::Shape(_) => ToolType::Shape,
 		ToolMessage::Ellipse(_) => ToolType::Ellipse,
 		ToolMessage::Polygon(_) => ToolType::Polygon,
 		ToolMessage::Text(_) => ToolType::Text,
@@ -436,7 +436,7 @@ pub fn tool_type_to_activate_tool_message(tool_type: ToolType) -> ToolMessageDis
 		ToolType::Freehand => ToolMessageDiscriminant::ActivateToolFreehand,
 		ToolType::Spline => ToolMessageDiscriminant::ActivateToolSpline,
 		ToolType::Line => ToolMessageDiscriminant::ActivateToolLine,
-		ToolType::Rectangle => ToolMessageDiscriminant::ActivateToolRectangle,
+		ToolType::Shape => ToolMessageDiscriminant::ActivateToolShape,
 		ToolType::Ellipse => ToolMessageDiscriminant::ActivateToolEllipse,
 		ToolType::Polygon => ToolMessageDiscriminant::ActivateToolPolygon,
 		ToolType::Text => ToolMessageDiscriminant::ActivateToolText,
